--- examples/unifews/config/cora.json
@@ -1,16 +1,0 @@
-{
-    "data": "cora",
-    "path": "./data/",
-    "algo": "gcn_unifews",
-    "epochs": 200,
-    "patience": 20,
-    "lr": 0.001,
-    "weight_decay": 1e-5,
-    "layer": 2,
-    "hidden": 512,
-    "dropout": 0.5,
-    "thr_a": 0.5,
-    "thr_w": 0.5,
-    "inductive": false,
-    "multil": false
-}
