--- examples/unifews/data/cora/attribute.txt
@@ -1,2 +1,0 @@
-n=2485
-m=12623
