+++ examples/unifews/run_mlp.py
@@ -1,84 +1,24 @@
-import os
-if 'TL_BACKEND' not in os.environ:
-    os.environ['TL_BACKEND'] = 'torch'
-
 import gc
 import random
 import argparse
 import numpy as np
-import ptflops
+
+try:
+    import ptflops
+    HAS_PTFLOPS = True
+except ImportError:
+    HAS_PTFLOPS = False
 
 import tensorlayerx as tlx
 from tensorlayerx import nn
 from tensorlayerx.dataflow import Dataset, DataLoader
 from tensorlayerx.model import WithLoss, TrainOneStep
 
-
-class PureTLXAdam:
-    def __init__(self, lr=0.001, weight_decay=0.0):
-        self.lr = lr
-        self.weight_decay = weight_decay
-        self.m = []
-        self.v = []
-        self.t = 0
-        self.beta1 = 0.9
-        self.beta2 = 0.999
-        self.epsilon = 1e-7
-
-    def gradient(self, loss, weights):
-        
-        if hasattr(loss, 'backward'):
-            loss.backward()
-        
-        grads = []
-        for w in weights:
-            grad = w.grad if hasattr(w, 'grad') else tlx.zeros_like(w)
-            grads.append(grad)
-        return grads
-
-    def apply_gradients(self, grads_and_vars):
-       
-        grads_and_vars = list(grads_and_vars)
-        if not grads_and_vars:
-            return
-        
-        self.t += 1
-        bias_correction1 = 1.0 - self.beta1**self.t
-        bias_correction2 = 1.0 - self.beta2**self.t
-        step_size = self.lr * (bias_correction2**0.5) / bias_correction1
-       
-        if len(self.m) != len(grads_and_vars):
-            self.m = [tlx.zeros_like(v) for g, v in grads_and_vars]
-            self.v = [tlx.zeros_like(v) for g, v in grads_and_vars]
-        
-        for i, (g, p) in enumerate(grads_and_vars):
-            if g is None:
-                continue
-            
-            
-            if self.weight_decay != 0:
-                g = g + self.weight_decay * p
-            
-           
-            self.m[i] = self.beta1 * self.m[i] + (1.0 - self.beta1) * g
-            self.v[i] = self.beta2 * self.v[i] + (1.0 - self.beta2) * (g * g)
-            
-
-            update = step_size * (self.m[i] / (tlx.sqrt(self.v[i]) + self.epsilon))
-            
-          
-            p.data -= update
-            
-            if hasattr(p, 'grad') and p.grad is not None:
-                p.grad = None
-# =====================================================================
-
-
-# =====================================================================
-
-from gammagl.utils.logger_gamma import Logger, ModelLogger, prepare_opt
-from gammagl.utils.loader_gamma import load_embedding
-import gammagl.utils.metric_gamma as metric
+import sys
+sys.path.append("/root/GammaGL")
+from gammagl.utils.logger_unifews import Logger, ModelLogger
+from examples.unifews.loader import load_embedding
+import gammagl.utils.metric_unifews as metric
 from gammagl.models.gnn_unifews import flops_modules_dict
 import gammagl.models.mlp_unifews as models
 
@@ -88,21 +28,33 @@ np.set_printoptions(linewidth=160, edgeitems=5, threshold=20,
 # ========== Training settings ==========
 parser = argparse.ArgumentParser()
 parser.add_argument('-f', '--seed', type=int, default=11)
-parser.add_argument('-v', '--dev', type=int, default=1)
-parser.add_argument('-c', '--config', type=str, default='./config/cora_mb.json')
-parser.add_argument('-m', '--algo', type=str, default=None)
+parser.add_argument('-v', '--dev', type=int, default=0)
+parser.add_argument('-m', '--algo', type=str, default='sgc')
 parser.add_argument('-n', '--suffix', type=str, default='')
-parser.add_argument('-a', '--thr_a', type=float, default=None)
-parser.add_argument('-w', '--thr_w', type=float, default=None)
-parser.add_argument('-l', '--layer', type=int, default=None)
-parser.add_argument('-p', '--hop', type=int, default=None)
-args = prepare_opt(parser)
+parser.add_argument('-a', '--thr_a', type=float, default=0.0001)
+parser.add_argument('-w', '--thr_w', type=float, default=0.1)
+parser.add_argument('-l', '--layer', type=int, default=2)
+parser.add_argument('-p', '--hop', type=int, default=2)
+parser.add_argument('--data', type=str, default='cora')
+parser.add_argument('--path', type=str, default='./data/')
+parser.add_argument('--epochs', type=int, default=200)
+parser.add_argument('--patience', type=int, default=20)
+parser.add_argument('--lr', type=float, default=0.001)
+parser.add_argument('--weight_decay', type=float, default=1e-5)
+parser.add_argument('--hidden', type=int, default=512)
+parser.add_argument('--dropout', type=float, default=0.5)
+parser.add_argument('--batch', type=int, default=64)
+parser.add_argument('--inductive', action='store_true', default=False)
+parser.add_argument('--multil', action='store_true', default=False)
+args = parser.parse_args()
+
+# Construct chn config (previously loaded from JSON)
+args.chn = {'hop': args.hop, 'delta': args.thr_a, 'alpha': 0.0, 'rrz': 0.5}
 
 num_thread = 0 if args.data in ['cora', 'citeseer', 'pubmed'] else 8
 random.seed(args.seed)
 np.random.seed(args.seed)
 tlx.set_seed(args.seed)
-device = f'cuda:{args.dev}' if args.dev >= 0 else 'cpu'
 
 if '_' not in args.algo: args.thr_a, args.thr_w = 0.0, 0.0
 args.chn['delta'] = args.thr_a
@@ -112,7 +64,7 @@ flag_run = f"{args.seed}-{args.thr_a:.1e}-{args.thr_w:.1e}"
 logger = Logger(args.data, args.algo, flag_run=flag_run)
 logger.save_opt(args)
 model_logger = ModelLogger(logger, patience=args.patience, cmp='max',
-                           prefix=f'model{args.suffix}', storage='state_gpu')
+                           prefix=f'model{args.suffix}', storage='state')
 stopwatch = metric.Stopwatch()
 
 # ========== Data Load ==========
@@ -127,13 +79,9 @@ model = models.MLP_unifews(
     thr_w=args.thr_w, dropout=args.dropout, layer=args.algo,
 )
 
-# 1. LAZY INIT FIX: Trigger model build using a CPU tensor first!
-dummy_x_cpu = tlx.convert_to_tensor(np.random.randn(1, nfeat), dtype=tlx.float32)
-_ = model(dummy_x_cpu, edge_idx=None)
-
-# 2. Now that weights are built, move the model to the target device
-if hasattr(model, 'to'):
-    model.to(device)
+# Trigger model build with a dummy forward pass
+dummy_x = tlx.convert_to_tensor(np.random.randn(1, nfeat), dtype=tlx.float32)
+_ = model(dummy_x, edge_idx=None)
 model.reset_parameters()
 
 if logger.lvl_config > 1:
@@ -145,11 +93,8 @@ loss_fn = tlx.losses.sigmoid_cross_entropy_with_logits if args.multil else tlx.l
 net_with_loss = WithLoss(model, loss_fn)
 
 train_weights = model.trainable_weights
-if not train_weights and hasattr(model, 'parameters'):
-    train_weights = [p for p in model.parameters() if p.requires_grad]
 
-# Use our Custom TLX Optimizer and standard TrainOneStep
-optimizer = PureTLXAdam(lr=args.lr, weight_decay=args.weight_decay)
+optimizer = tlx.optimizers.Adam(lr=args.lr, weight_decay=args.weight_decay)
 train_one_step = TrainOneStep(net_with_loss, optimizer, train_weights)
 
 best_val_acc = 0.0
@@ -176,9 +121,6 @@ def train(epoch, ld=loader_train, verbose=False):
     stopwatch.reset()
 
     for it, (x, y) in enumerate(ld):
-        x = tlx.convert_to_tensor(x, device=device)
-        y = tlx.convert_to_tensor(y, device=device)
-
         if it == 0:
             model.set_scheme('pruneall' if epoch < args.epochs // 2 else 'pruneinc', 'pruneall')
         else:
@@ -193,10 +135,6 @@ def train(epoch, ld=loader_train, verbose=False):
     return np.mean(loss_list), stopwatch.time
 
 # ========== Eval Func ==========
-def to_numpy(t):
-    # Use the native TLX converter - it handles GPU->CPU and Detaching automatically
-    return tlx.convert_to_numpy(t)
-
 def eval(ld, verbose=False):
     model.eval()
     model.set_scheme('keep', 'keep')
@@ -204,32 +142,23 @@ def eval(ld, verbose=False):
     stopwatch.reset()
 
     for x, y in ld:
-        # 1. Move input data to the GPU for inference
-        x = tlx.convert_to_tensor(x, device=device)
-        y = tlx.convert_to_tensor(y, device=device)
-
         stopwatch.start()
         output = model(x, edge_idx=None)
         stopwatch.pause()
 
-        # 2. Get predictions (still on GPU)
         if args.multil:
             output = tlx.where(output > 0, 1, 0)
         else:
             output = tlx.argmax(output, axis=1)
-        
-        # 3. FIX: Move to CPU but KEEP as Tensors
-        # This allows metric_gamma.py to use tlx.reshape/tlx.cast without crashing,
-        # and ensures it won't trigger the "cuda:1 to numpy" error later.
-        y_cpu = tlx.convert_to_tensor(y, device='cpu')
-        output_cpu = tlx.convert_to_tensor(output, device='cpu')
-        
-        calc.update(y_cpu, output_cpu)
+
+        calc.update(y, output)
 
     return calc.compute('micro'), stopwatch.time, None, None
 
 # ========== FLOPs Func ==========
 def cal_flops(ld, verbose=False):
+    if not HAS_PTFLOPS:
+        return 0.0, 0.0
     model.eval()
     model.set_scheme('keep', 'keep')
     macs, nparam = ptflops.get_model_complexity_info(
@@ -272,9 +201,7 @@ for epoch in range(1, args.epochs + 1):
         epoch_conv = epoch - model_logger.patience
 
 # ========== Test ==========
-model = model_logger.load()
-if hasattr(model, 'to'):
-    model.to(device)
+#model = model_logger.load()
 acc_test, time_test, _, _ = eval(loader_test)
 macs_test, _ = cal_flops(loader_val)
 macs_test *= len(idx['test'])
