+++ gammagl/utils/loader_unifews.py
@@ -1,0 +1,222 @@
+import os.path as osp
+import sys
+import gc
+import copy
+import os
+from dotmap import DotMap
+import numpy as np
+import scipy.sparse as sp
+
+import tensorlayerx as tlx
+from gammagl.data import Graph
+from typing import Any, Callable, List, Optional, Union, Sequence
+
+def stochastic_blockmodel_graph(block_sizes, edge_probs, directed=False):
+    N = sum(block_sizes)
+    edges = []
+    for i, size_i in enumerate(block_sizes):
+        start_i = sum(block_sizes[:i])
+        end_i = start_i + size_i
+        for j, size_j in enumerate(block_sizes):
+            start_j = sum(block_sizes[:j])
+            end_j = start_j + size_j
+            p = edge_probs[i, j] if isinstance(edge_probs, np.ndarray) else edge_probs[i][j]
+            if directed:
+                idx_i = np.random.choice(np.arange(start_i, end_i), size=int(size_i * size_j * p))
+                idx_j = np.random.choice(np.arange(start_j, end_j), size=int(size_i * size_j * p))
+            else:
+                if i > j:
+                    continue
+                n_edges = int(size_i * size_j * p)
+                idx_i, idx_j = np.random.randint(start_i, end_i, n_edges), np.random.randint(start_j, end_j, n_edges)
+                if i == j:
+                    mask = idx_i < idx_j
+                    idx_i, idx_j = idx_i[mask], idx_j[mask]
+            if len(idx_i) > 0:
+                edges.append(np.stack([idx_i, idx_j]))
+    edge_index = np.concatenate(edges, axis=1)
+    if not directed:
+        edge_index = np.concatenate([edge_index, edge_index[::-1]], axis=1)
+    return edge_index
+
+def to_scipy_sparse_matrix(edge_index, num_nodes=None):
+    
+    if num_nodes is None:
+        num_nodes = edge_index.max() + 1
+    return sp.coo_matrix((np.ones(edge_index.shape[1]), (edge_index[0], edge_index[1])),
+                         shape=(num_nodes, num_nodes))
+
+class DummySBM:
+    def __init__(self, edge_index, num_nodes):
+        self.edge_index = edge_index
+        self.num_nodes = num_nodes
+
+from pathlib import Path
+ROOT = Path(__file__).parent.parent.parent
+sys.path.append(str(ROOT))
+from .gen_cat import gencat, feature_extraction
+from .data_processor import DataProcess, DataProcess_inductive, matstd_clip
+
+np.set_printoptions(linewidth=160, edgeitems=5, threshold=20,
+                    formatter=dict(float=lambda x: "% 9.3e" % x))
+
+
+def to_list(value: Any) -> Sequence:
+    if isinstance(value, Sequence) and not isinstance(value, str):
+        return value
+    else:
+        return [value]
+
+
+# ===================== 彻底删除废弃的GammaGL数据集继承，纯手动实现 =====================
+def sbm_mixture_of_Gaussians(n, n_features, sizes, probs, std_, means, nodes_per_mean):
+    
+    edge_index = stochastic_blockmodel_graph(sizes, probs, directed=False)
+    g_ = DummySBM(edge_index=edge_index, num_nodes=sum(sizes))
+
+    A = to_scipy_sparse_matrix(g_.edge_index, num_nodes=g_.num_nodes)
+    A.setdiag(A.diagonal() + 1)
+    d_mat = np.sum(A, axis=0)
+    dinv_mat = 1/d_mat
+    dinv = dinv_mat.tolist()[0]
+    Dinv = sp.diags(dinv, 0)
+
+    X = np.zeros((2*n,n_features))
+    for ct, mean_ in enumerate(means):
+        X[nodes_per_mean[ct],:] = std_ * np.random.randn(len(nodes_per_mean[ct]), n_features) + mean_
+
+    data = np.zeros((2*n,n_features+1))
+    data[:,:-1] = Dinv@A@X
+    data[:,-1]  = np.ones(2*n)
+    g_.x = tlx.convert_to_tensor(data, dtype=tlx.float32)
+
+    y = np.zeros(2*n)
+    y[0:n] = 1
+    g_.y = tlx.convert_to_tensor(y, dtype=tlx.float32)
+
+    return g_
+
+
+def load_csbm(datastr: str, datapath: str="./data/",
+                   inductive: bool=False, multil: bool=False,
+                   seed: int=0, **kwargs):
+    
+    n = 500
+    n_features = 127
+    q = 0.1
+    _, p, mu = datastr.split('-')
+    p, mu = float(p), float(mu)
+
+    probs = [[p, q], [q, p]]
+    sizes = [n, n]
+    std_ = mu
+    means = [100, -100]
+    nodes_per_mean = [list(range(n)),list(range(n,2*n))]
+    
+    g = sbm_mixture_of_Gaussians(n, n_features, sizes, probs, std_, means, nodes_per_mean)
+
+    idx_rnd = np.random.permutation(2*n)
+    adj = {'train': g.edge_index,
+           'test':  g.edge_index}
+    feat = {'train': tlx.convert_to_tensor(g.x, tlx.float32),
+           'test':  tlx.convert_to_tensor(g.x, tlx.float32)}
+    idx = {'train': tlx.convert_to_tensor(idx_rnd[:int(0.6*2*n)], tlx.int64),
+           'val':   tlx.convert_to_tensor(idx_rnd[int(0.6*2*n):int(0.8*2*n)], tlx.int64),
+           'test':  tlx.convert_to_tensor(idx_rnd[int(0.8*2*n):], tlx.int64)}
+    y = tlx.cast(g.y, tlx.int64)
+    labels = {'train': y[idx['train']],
+              'val':   y[idx['val']],
+              'test':  y[idx['test']]}
+    nfeat = n_features + 1
+    nclass = 2
+    if seed >= 15:
+        print(g)
+    return adj, feat, labels, idx, nfeat, nclass
+
+
+def load_gencat(datastr: str, datapath: str="./data/",
+                   inductive: bool=False, multil: bool=False,
+                   seed: int=0, **kwargs):
+    dp = DataProcess('cora', path="./data/", seed=0)
+    dp.input(['adjnpz', 'labels', 'attr_matrix'])
+    M, D, class_size, H, node_degree = feature_extraction(dp.adj_matrix, dp.attr_matrix, dp.labels.tolist())
+
+    _, p, omega = datastr.split('-')
+    p, omega = float(p), float(omega)
+    adj, feat, labels = gencat(M, D, H, n=1000, m=5000, p=p, omega=omega)
+    dp.adj_matrix = adj
+    dp.attr_matrix = feat
+    dp.labels = np.array(labels)
+
+    dp.calculate(['idx_train'])
+    idx = {'train': tlx.convert_to_tensor(dp.idx_train, tlx.int64),
+           'val':   tlx.convert_to_tensor(dp.idx_val, tlx.int64),
+           'test':  tlx.convert_to_tensor(dp.idx_test, tlx.int64)}
+    labels = tlx.convert_to_tensor(dp.labels.flatten(), tlx.int64)
+    labels = {'train': labels[idx['train']],
+              'val':   labels[idx['val']],
+              'test':  labels[idx['test']]}
+    dp.calculate(['edge_idx'])
+    adj = {'test':  tlx.convert_to_tensor(dp.edge_idx, tlx.int64),
+           'train': tlx.convert_to_tensor(dp.edge_idx, tlx.int64)}
+    feat = {'test': tlx.convert_to_tensor(dp.attr_matrix, tlx.float32),
+            'train': tlx.convert_to_tensor(dp.attr_matrix, tlx.float32)}
+    n, m = dp.n, dp.m
+    nfeat, nclass = dp.nfeat, dp.nclass
+    if seed >= 15:
+        print(dp)
+    return adj, feat, labels, idx, nfeat, nclass
+
+def load_edgelist(datastr: str, datapath: str="./data/",
+                   inductive: bool=False, multil: bool=False,
+                   seed: int=0, **kwargs):
+    if datastr.startswith('csbm'):
+        return load_csbm(datastr, datapath, inductive, multil, seed, **kwargs)
+    elif datastr.startswith('gencat'):
+        return load_gencat(datastr, datapath, inductive, multil, seed, **kwargs)
+
+    dp = DataProcess(datastr, path=datapath, seed=seed)
+    dp.input(['adjnpz', 'labels', 'attr_matrix'])
+    if inductive:
+        dpi = DataProcess_inductive(datastr, path=datapath, seed=seed)
+        dpi.input(['adjnpz', 'attr_matrix'])
+    else:
+        dpi = dp
+
+    if (datastr.startswith('cora') or datastr.startswith('citeseer') or datastr.startswith('pubmed')):
+        dp.calculate(['idx_train'])
+    else:
+        dp.input(['idx_train', 'idx_val', 'idx_test'])
+    idx = {'train': tlx.convert_to_tensor(dp.idx_train, tlx.int64),
+           'val':   tlx.convert_to_tensor(dp.idx_val, tlx.int64),
+           'test':  tlx.convert_to_tensor(dp.idx_test, tlx.int64)}
+
+    if multil:
+        dp.calculate(['labels_oh'])
+        dp.labels_oh[dp.labels_oh < 0] = 0
+        labels = tlx.convert_to_tensor(dp.labels_oh, tlx.float32)
+    else:
+        dp.labels[dp.labels < 0] = 0
+        labels = tlx.convert_to_tensor(dp.labels.flatten(), tlx.int64)
+    labels = {'train': labels[idx['train']],
+              'val':   labels[idx['val']],
+              'test':  labels[idx['test']]}
+
+    dp.calculate(['edge_idx'])
+    adj = {'test':  tlx.convert_to_tensor(dp.edge_idx, tlx.int64)}
+    if inductive:
+        dpi.calculate(['edge_idx'])
+        adj['train'] = tlx.convert_to_tensor(dpi.edge_idx, tlx.int64)
+    else:
+        adj['train'] = adj['test']
+
+    feat = dp.attr_matrix
+    feati = dpi.attr_matrix if inductive else feat
+    feat = {'test': tlx.convert_to_tensor(feat, tlx.float32)}
+    feat['train'] = tlx.convert_to_tensor(feati, tlx.float32)
+
+    n, m = dp.n, dp.m
+    nfeat, nclass = dp.nfeat, dp.nclass
+    if seed >= 15:
+        print(dp)
+    return adj, feat, labels, idx, nfeat, nclass
